feat: allow mount zally-web-ui on a specific path
* add logger options
* add refactor

close #3
