+++ .npmignore
@@ -5,6 +5,7 @@ jest.json
 
 __tests__
 server.js
+server.mount.js
 
 webpack.config.js
 target
