--- src/client/public/index.html
@@ -1,14 +1,0 @@
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Zally API Linter WEB UI</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="https://fonts.googleapis.com/css?family=Ubuntu:300,500,300italic|Ubuntu+Mono" rel="stylesheet" type="text/css">
-  <link href="https://unpkg.com/dress-code@2.0.0/dist/css/dress-code.min.css" rel="stylesheet" type="text/css">
-  <script src="/env.js" type="text/javascript"></script>
-</head>
-<body>
-<div id="app"></div>
-<script src="/assets/bundle.js" type="text/javascript"></script>
-</body>
-</html>
