+++ src/server/views/index.pug
@@ -1,0 +1,13 @@
+html
+  head
+    meta(charset="utf-8")
+    title=headTitle
+    base(href=mountpath)
+    meta(name="viewport" content="width=device-width, initial-scale=1")
+    link(href="https://fonts.googleapis.com/css?family=Ubuntu:300,500,300italic|Ubuntu+Mono" rel="stylesheet" type="text/css")
+    link(href="https://unpkg.com/dress-code@2.0.0/dist/css/dress-code.min.css" rel="stylesheet" type="text/css")
+    script(src=mountpath + "env.js" type="text/javascript")
+  body
+    div(id="app")
+    script(src=mountpath + "assets/bundle.js" type="text/javascript")
+
